backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
redis
torch
torchvision
pillow
aioredis
pytest
httpx
alembic
supabase
celery
redis
requests
sentry-sdk
slowapi
locust
posthog
firebase-admin
Pillow
